moved usescases, tests
