+++ lib/actions/sendMessageToOih.js
@@ -1,7 +1,7 @@
 const { messages } = require('elasticio-node');
 const { EventBus, Event, RabbitMqTransport } = require('@openintegrationhub/event-bus');
 const bunyan = require('bunyan');
-const request = require('request-promise');
+
 
 /**
  * Executes the action's logic by ...
@@ -34,8 +34,6 @@ async function processAction(msg, cfg) {
     meta,
   };
 
-  console.log(`Send a new message to OIH with content ${JSON.stringify(messageWithMeta)}`);
-
   const transport = new RabbitMqTransport({ rabbitmqUri: 'amqp://guest:guest@localhost:5672', logger });
   const eventBus = new EventBus({ transport, logger, serviceName: 'my-service' });
   await eventBus.connect();
@@ -49,41 +47,8 @@ async function processAction(msg, cfg) {
   const result = await eventBus.publish(event);
   console.log(`Published new message to Eventbus with content ${JSON.stringify(result)}`);
 
+  console.log(`Send a new message to OIH with content ${JSON.stringify(messageWithMeta)}`);
   return messages.newMessageWithBody(messageWithMeta);
 }
 
-async function getDomains() {
-  const requestOptions = {
-    method: 'GET',
-    uri: 'http://metadata.openintegrationhub.com/api/v1/domains',
-    json: true,
-    headers: {
-      Authorization: 'Bearer xxxxxx',
-    },
-  };
-
-  const domains = await request.get(requestOptions);
-  const domainNames = domains.map(element => element.name);
-  console.log(JSON.stringify(domainNames));
-  return domainNames;
-}
-
-async function getModels(cfg) {
-  const requestOptions = {
-    method: 'GET',
-    uri: `http://metadata.openintegrationhub.com/api/v1/domains/${cfg.domain}/schemas`,
-    json: true,
-    headers: {
-      Authorization: 'Bearer xxxxxx',
-    },
-  };
-
-  const models = await request.get(requestOptions);
-  const modelNames = models.map(element => element.name);
-  console.log(JSON.stringify(modelNames));
-  return modelNames;
-}
-
 exports.process = processAction;
-exports.domains = getDomains;
-exports.models = getModels;
